set _NET_WM_WINDOW_TYPE_DOCK atom when statusbar
Setting this atom is enough for i3 to treat the window as a
statusbar. i3 will automatically populate the bar on each workspace
screen and create the correct struts.

Other property atoms will need to be set for compatibility with other
Window managers that are not as proactive.
For a list of property atoms that may need to be set see for example:
https://github.com/LemonBoy/bar/blob/master/lemonbar.c